feat: SJIP-128 cleanup
